+++ tuning_fork_samples/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.1
+
+* The README credits Sonic Pi for the collection.
+
 ## 0.1.0
 
 First release.